refactor: move mark interfaces to common package
