+++ app/src/main/kotlin/kr/ohyung/navigation/common/Api.kt
@@ -1,6 +1,6 @@
 /*
  * Created by Lee Oh Hyoung on 2020/08/06 .. 
  */
-package kr.ohyung.navigation
+package kr.ohyung.navigation.common
 
-interface UiState
+interface Api
